@@ -4,44 +4,51 @@
   "description": "client-side the-messenger",
   "main": "index.js",
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
+    "start": "craco start",
+    "build": "craco build",
+    "test": "craco test --env=jsdom",
     "eject": "react-scripts eject"
   },
   "proxy": "http://localhost:5000",
   "devDependencies": {
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^7.2.0",
-    "eslint-plugin-prettier": "^3.3.1",
-    "eslint-plugin-react": "^7.21.4",
-    "prettier": "^2.2.1",
+    "eslint": "^8.2.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "prettier": "^2.5.1",
     "redux-devtools": "^3.7.0"
   },
   "author": "PLFJ",
   "license": "ISC",
   "dependencies": {
-    "@react-pdf/renderer": "^1.6.11",
-    "axios": "^0.21.0",
-    "bootstrap": "^4.5.2",
-    "core-js": "^3.6.5",
+    "@babel/eslint-parser": "^7.16.5",
+    "@craco/craco": "^6.4.3",
+    "@react-pdf/renderer": "^2.1.0",
+    "axios": "^0.24.0",
+    "bootstrap": "^5.1.3",
+    "core-js": "^3.20.3",
     "emoji-mart": "^3.0.1",
     "eslint-config-node": "^4.1.0",
     "eslint-plugin-node": "^11.1.0",
-    "firebase": "^8.1.2",
-    "jquery": "^3.5.1",
-    "react": "^16.13.1",
-    "react-bootstrap": "^1.4.1",
-    "react-dom": "^16.13.1",
-    "react-redux": "^7.2.1",
-    "react-responsive": "^8.1.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^3.4.3",
-    "react-transition-group": "^4.4.1",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0",
+    "firebase": "^9.6.2",
+    "jquery": "^3.6.0",
+    "react": "^17.0.2",
+    "react-bootstrap": "^2.1.0",
+    "react-dom": "^17.0.2",
+    "react-redux": "^7.2.6",
+    "react-responsive": "^9.0.0-beta.6",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "^5.0.0",
+    "react-transition-group": "^4.4.2",
+    "redux": "^4.1.2",
+    "redux-thunk": "^2.4.1",
+    "source-map": "^0.7.3",
+    "svgo": "^2.8.0",
     "use-dynamic-refs": "^1.0.0",
-    "uuid": "^8.3.1"
+    "uuid": "^8.3.2"
   },
   "resolutions": {
     "dfa": "1.2.0"
