@@ -1,6 +1,5 @@
 import React from 'react'
-import './SelectView.css'
-import { Form, Row } from 'react-bootstrap'
+import { Form } from 'react-bootstrap'
 import { connect } from 'react-redux'
 import { changeModus } from '../../redux/actions/user'
 
@@ -24,72 +23,48 @@ export function SelectView(props) {
 
   // ----------------------------------------------------------------------------------------------------------
   return (
-    <Form as={Row} id={props.id}>
-      <div id="size">
-        <Form.Group as={Row}>
-          <Form.Check
-            type="radio"
-            id="auto"
-            name="modi"
-            label=""
-            onChange={(e) => handleChange(e)}
-            checked={props.user.modus === 'auto' ? true : false}
-          />
-          <Form.Label id="label-view">auto</Form.Label>
-        </Form.Group>
-      </div>
-      <div id="size">
-        <Form.Group as={Row}>
-          <Form.Check
-            type="radio"
-            id="desktop"
-            name="modi"
-            label=""
-            onChange={(e) => handleChange(e)}
-            checked={props.user.modus === 'desktop' ? true : false}
-          />
-          <Form.Label
-            id="label-view"
-            style={
-              window.innerWidth <= 979 ? { color: 'grey' } : { color: 'white' }
-            }
-          >
-            Desktop
-          </Form.Label>
-        </Form.Group>
-      </div>
-      <div id="size">
-        <Form.Group as={Row}>
-          <Form.Check
-            type="radio"
-            id="tablet"
-            name="modi"
-            label=""
-            onChange={(e) => handleChange(e)}
-            checked={props.user.modus === 'tablet' ? true : false}
-          />
-          <Form.Label
-            id="label-view"
-            style={
-              window.innerWidth <= 767 ? { color: 'grey' } : { color: 'white' }
-            }
-          >
-            Tablet
-          </Form.Label>
-        </Form.Group>
-      </div>
-      <div id="size">
-        <Form.Group as={Row}>
-          <Form.Check
-            type="radio"
-            id="mobile"
-            name="modi"
-            label=""
-            onChange={(e) => handleChange(e)}
-            checked={props.user.modus === 'mobile' ? true : false}
-          />
-          <Form.Label id="label-view">Mobile</Form.Label>
-        </Form.Group>
+    <Form id={props.id}>
+      <div>
+        <Form.Check
+          style={{ fontSize: '12px' }}
+          inline
+          label="auto"
+          name="modi"
+          type="radio"
+          id="auto"
+          onChange={(e) => handleChange(e)}
+          checked={props.user.modus === 'auto' ? true : false}
+        />
+        <Form.Check
+          style={{ fontSize: '12px' }}
+          inline
+          label="desktop"
+          name="modi"
+          type="radio"
+          id="desktop"
+          onChange={(e) => handleChange(e)}
+          checked={props.user.modus === 'desktop' ? true : false}
+        />
+        <Form.Check
+          style={{ fontSize: '12px' }}
+          inline
+          label="tablet"
+          name="modi"
+          type="radio"
+          id="tablet"
+          onChange={(e) => handleChange(e)}
+          checked={props.user.modus === 'tablet' ? true : false}
+        />
+        <Form.Check
+          style={{ fontSize: '12px' }}
+          inline
+          label="mobile"
+          name="modi"
+          type="radio"
+          id="mobile"
+          onChange={(e) => handleChange(e)}
+          checked={props.user.modus === 'mobile' ? true : false}
+        />
       </div>
     </Form>
   )
@@ -106,3 +81,69 @@ const mapActionsToProps = {
 }
 
 export default connect(mapStateToProps, mapActionsToProps)(SelectView)
+
+/*
+<Form as={Row} id={props.id}>
+      <FormGroup as={Row} id="size">
+        <Radio
+          id="auto"
+          name="modi"
+          onChange={(e) => handleChange(e)}
+          checked={props.user.modus === 'auto' ? true : false}
+          inline
+        >
+          auto
+        </Radio>
+      </FormGroup>
+
+      <Form.Group as={Row} id="size">
+        <Form.Check
+          type="radio"
+          id="desktop"
+          name="modi"
+          label=""
+          onChange={(e) => handleChange(e)}
+          checked={props.user.modus === 'desktop' ? true : false}
+        />
+        <Form.Label
+          id="label-view"
+          style={
+            window.innerWidth <= 979 ? { color: 'grey' } : { color: 'white' }
+          }
+        >
+          Desktop
+        </Form.Label>
+      </Form.Group>
+
+      <Form.Group as={Row} id="size">
+        <Form.Check
+          type="radio"
+          id="tablet"
+          name="modi"
+          label=""
+          onChange={(e) => handleChange(e)}
+          checked={props.user.modus === 'tablet' ? true : false}
+        />
+        <Form.Label
+          id="label-view"
+          style={
+            window.innerWidth <= 767 ? { color: 'grey' } : { color: 'white' }
+          }
+        >
+          Tablet
+        </Form.Label>
+      </Form.Group>
+
+      <Form.Group as={Row} id="size">
+        <Form.Check
+          type="radio"
+          id="mobile"
+          name="modi"
+          label=""
+          onChange={(e) => handleChange(e)}
+          checked={props.user.modus === 'mobile' ? true : false}
+        />
+        <Form.Label id="label-view">Mobile</Form.Label>
+      </Form.Group>
+    </Form>
+    */
